+++ sdk/src/main/java/com/eyeson/sdk/events/EyesonEventListener.kt
@@ -11,6 +11,7 @@ import java.util.*
 abstract class EyesonEventListener {
     open fun onPermissionsNeeded(neededPermissions: List<NeededPermissions>) {}
     open fun onMeetingJoining(
+        accessKey: String,
         name: String,
         startedAt: Date,
         user: UserInfo,
@@ -49,7 +50,7 @@ abstract class EyesonEventListener {
 
     open fun onUserJoinedMeeting(users: List<UserInfo>) {}
     open fun onUserLeftMeeting(users: List<UserInfo>) {}
-    open fun onUserListUpdate(users: List<UserInfo>) {}
+    open fun onUserListUpdate(users: List<UserInfo>, playbackPlayIds:List<String>) {}
     open fun onVoiceActivity(user: UserInfo, active: Boolean) {}
 
     open fun onChatMessageReceived(user: UserInfo, message: String, timestamp: Date) {}