+++ sdk/src/main/java/com/eyeson/sdk/model/local/sepp/MemberListUpdate.kt
@@ -5,5 +5,6 @@ import com.eyeson.sdk.model.local.base.LocalBaseCommand
 internal data class MemberListUpdate(
     val added: List<String>,
     val deleted: List<String>,
-    val memberCountAfterUpdate: Int
+    val memberCountAfterUpdate: Int,
+    val mediaPlayIds:List<String>
 ) : LocalBaseCommand